+++ core/api/golang/kurtosis_core_rpc_api_bindings/api_container_service_grpc.pb.go
@@ -57,7 +57,7 @@ type ApiContainerServiceClient interface {
 	// Uploads a files artifact to the Kurtosis File System.
 	UploadFilesArtifact(ctx context.Context, in *UploadFilesArtifactArgs, opts ...grpc.CallOption) (*UploadFilesArtifactResponse, error)
 	// Tells the API container to download a files artifact from the web to the Kurtosis File System
-	DownloadFilesArtifact(ctx context.Context, in *DownloadFilesArtifactArgs, opts ...grpc.CallOption) (*DownloadFilesArtifactResponse, error)
+	StoreWebFilesArtifact(ctx context.Context, in *StoreWebFilesArtifactArgs, opts ...grpc.CallOption) (*StoreWebFilesArtifactResponse, error)
 }
 
 type apiContainerServiceClient struct {
@@ -212,9 +212,9 @@ func (c *apiContainerServiceClient) UploadFilesArtifact(ctx context.Context, in 
 	return out, nil
 }
 
-func (c *apiContainerServiceClient) DownloadFilesArtifact(ctx context.Context, in *DownloadFilesArtifactArgs, opts ...grpc.CallOption) (*DownloadFilesArtifactResponse, error) {
-	out := new(DownloadFilesArtifactResponse)
-	err := c.cc.Invoke(ctx, "/api_container_api.ApiContainerService/DownloadFilesArtifact", in, out, opts...)
+func (c *apiContainerServiceClient) StoreWebFilesArtifact(ctx context.Context, in *StoreWebFilesArtifactArgs, opts ...grpc.CallOption) (*StoreWebFilesArtifactResponse, error) {
+	out := new(StoreWebFilesArtifactResponse)
+	err := c.cc.Invoke(ctx, "/api_container_api.ApiContainerService/StoreWebFilesArtifact", in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -259,7 +259,7 @@ type ApiContainerServiceServer interface {
 	// Uploads a files artifact to the Kurtosis File System.
 	UploadFilesArtifact(context.Context, *UploadFilesArtifactArgs) (*UploadFilesArtifactResponse, error)
 	// Tells the API container to download a files artifact from the web to the Kurtosis File System
-	DownloadFilesArtifact(context.Context, *DownloadFilesArtifactArgs) (*DownloadFilesArtifactResponse, error)
+	StoreWebFilesArtifact(context.Context, *StoreWebFilesArtifactArgs) (*StoreWebFilesArtifactResponse, error)
 	mustEmbedUnimplementedApiContainerServiceServer()
 }
 
@@ -315,8 +315,8 @@ func (UnimplementedApiContainerServiceServer) GetModules(context.Context, *empty
 func (UnimplementedApiContainerServiceServer) UploadFilesArtifact(context.Context, *UploadFilesArtifactArgs) (*UploadFilesArtifactResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UploadFilesArtifact not implemented")
 }
-func (UnimplementedApiContainerServiceServer) DownloadFilesArtifact(context.Context, *DownloadFilesArtifactArgs) (*DownloadFilesArtifactResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method DownloadFilesArtifact not implemented")
+func (UnimplementedApiContainerServiceServer) StoreWebFilesArtifact(context.Context, *StoreWebFilesArtifactArgs) (*StoreWebFilesArtifactResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method StoreWebFilesArtifact not implemented")
 }
 func (UnimplementedApiContainerServiceServer) mustEmbedUnimplementedApiContainerServiceServer() {}
 
@@ -619,20 +619,20 @@ func _ApiContainerService_UploadFilesArtifact_Handler(srv interface{}, ctx conte
 	return interceptor(ctx, in, info, handler)
 }
 
-func _ApiContainerService_DownloadFilesArtifact_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(DownloadFilesArtifactArgs)
+func _ApiContainerService_StoreWebFilesArtifact_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(StoreWebFilesArtifactArgs)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(ApiContainerServiceServer).DownloadFilesArtifact(ctx, in)
+		return srv.(ApiContainerServiceServer).StoreWebFilesArtifact(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/api_container_api.ApiContainerService/DownloadFilesArtifact",
+		FullMethod: "/api_container_api.ApiContainerService/StoreWebFilesArtifact",
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(ApiContainerServiceServer).DownloadFilesArtifact(ctx, req.(*DownloadFilesArtifactArgs))
+		return srv.(ApiContainerServiceServer).StoreWebFilesArtifact(ctx, req.(*StoreWebFilesArtifactArgs))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -709,8 +709,8 @@ var ApiContainerService_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _ApiContainerService_UploadFilesArtifact_Handler,
 		},
 		{
-			MethodName: "DownloadFilesArtifact",
-			Handler:    _ApiContainerService_DownloadFilesArtifact_Handler,
+			MethodName: "StoreWebFilesArtifact",
+			Handler:    _ApiContainerService_StoreWebFilesArtifact_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},